backend/requirements.txt
Flask==2.3.2
gunicorn==20.1.0
requests==2.31.0
yfinance==0.2.25
pandas==2.2.2
numpy==1.26.2
ta==0.11.0
scikit-learn==1.3.2
joblib==1.3.2
python-dotenv==1.0.0
feedparser==6.0.10
firebase-admin==6.0.1
PyJWT==2.8.0
